conexion-backend

+++ src/hooks/useCalendarStore.js
@@ -1,5 +1,8 @@
 import { useDispatch, useSelector } from "react-redux"
-import { onAddNewEvent, onDeleteEvent, onSetActiveEvent, onUpdateEvent } from "../store";
+import { onAddNewEvent, onDeleteEvent, onLoadEvents, onSetActiveEvent, onUpdateEvent } from "../store";
+import { calendarApi } from "../api";
+import { convertEventsToDateEvents } from "../helpers";
+import Swal from "sweetalert2";
 
 
 export const useCalendarStore = () => {
@@ -7,30 +10,68 @@ export const useCalendarStore = () => {
     const dispatch = useDispatch();
 
     const { events, activeEvent } = useSelector(state => state.calendar);
+    const { user } = useSelector(state => state.auth);
 
     const setActiveEvent = (calendarEvent) => {
         dispatch(onSetActiveEvent(calendarEvent));
     }
 
     const startSavingEvent = async (calendarEvent) => {
-        // TODO: llegar al backend
 
-        // Todo bien
-        if (calendarEvent._id) {
-            // Actualizando  
-            dispatch(onUpdateEvent({ ...calendarEvent }));
-        }
-        else {
+        try {
+            if (calendarEvent.id) {
+                // Actualizando  
+                await calendarApi.put(`/events/${calendarEvent.id}`, calendarEvent);
+
+                dispatch(onUpdateEvent({ ...calendarEvent }, user));
+
+                return;
+            }
+
             // Creando
-            dispatch(onAddNewEvent({ ...calendarEvent, _id: new Date().getTime() }));
+            const { data } = await calendarApi.post('/events', calendarEvent);
+
+            dispatch(onAddNewEvent({ ...calendarEvent, id: data.evento.id, user }));
+        } catch (error) {
+            console.log(error);
+            Swal.fire('Error al guardar', error.response.data.msg, 'error');
         }
     }
 
-    const startDeleteEvent = () => {
-        // TODO: llegar al backend
+    const startDeleteEvent = async (id) => {
 
+        try {
+            if (!!activeEvent) {
 
-        dispatch(onDeleteEvent());
+                await calendarApi.delete(`/events/${activeEvent.id}`);
+
+                dispatch(onDeleteEvent());
+
+                return;
+            }
+
+            Swal.fire('No hay ningún evento seleccionado', error.response.data.msg, 'error');
+
+            return;
+
+        } catch (error) {
+            console.log(error);
+            Swal.fire('Error al eliminar', error.response.data.msg, 'error');
+        }
+    }
+
+    const startLoadingEvents = async () => {
+        try {
+            const { data } = await calendarApi.get('/events');
+
+            const events = convertEventsToDateEvents(data.eventos);
+
+            dispatch(onLoadEvents(events));
+
+        } catch (error) {
+            console.log('Error cargando eventos');
+            console.log(error);
+        }
     }
 
     return {
@@ -41,7 +82,8 @@ export const useCalendarStore = () => {
 
         //* Methods
         setActiveEvent,
+        startDeleteEvent,
+        startLoadingEvents,
         startSavingEvent,
-        startDeleteEvent
     }
 }
